requirements.txt
Flask==2.3.2
pytest==7.3.2
pytest-mock==3.11.1
mypy==1.3.0
mypy-extensions==1.0.0
alembic==1.11.1
dataclasses-json==0.5.8
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
flask-restx==1.1.0
flask_sqlalchemy==3.0.4
sqlalchemy==2.0.16
python-dotenv==1.0.0
marshmallow==3.19.0
flask-marshmallow==0.15.0